# apispec.yaml
openapi: '3.0.3'
info:
    title: Bringeee-API
    version: v1.0
    description: Cargo Agency management system API spesification built with echo golang
components:
    securitySchemes:
        JWTAuthCustomer:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Use customer token for authorization'
        JWTAuthDriver:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Use driver token for authorization'
        JWTAuthAdmin:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Use admin token for authorization'
    schemas:
...
        pagination:
            type: object
            properties:
                page:
                    type: number
                    example: 1
                limit:
                    type: number
                    example: 50
                total_pages:
                    type: number
                    example: 7
                total_records:
                    type: number
                    example: 331
...
        success_response:
            type: object
            properties:
                status:
                    type: string
                    example: 'OK'
                error:
                    type: string
                    example: null
                links:
                    type: object
                    properties:
                        self:
                            type: string
                            example: 'somelink'
                data:
                    type: object
                    properties:
                        id:
                            type: number
                            example: 21
        error_validation_response:
            type: object
            properties:
                status:
                    type: string
                    example: 'ERROR'
                code:
                    type: number
                    example: 422
                error:
                    type: string
                    example: 'Form validation error'
                errors:
                    type: array
                    items:
                        type: object
                        properties:
                            name:
                                type: string
                                example: 'input_field'
                            error:
                                type: string
                                example: 'input_field is required'
                links:
                    type: object
                    properties:
                        self:
                            type: string
                            example: 'https://sandbox.something.com/somelink'
        error_unauthorized:
            type: object
            properties:
                status:
                    type: string
                    example: 'ERROR'
                code:
                    type: number
                    example: 401
                error:
                    type: string
                    example: 'Unauthorized user'
                links:
                    type: object
                    properties:
                        self:
                            type: string
                            example: 'https://sandbox.something.com/somelink'
        error_server_response:
            type: object
            properties:
                status:
                    type: string
                    example: 'ERROR'
                code:
                    type: number
                    example: 500
                error:
                    type: string
                    example: 'server error'
                links:
                    type: object
                    properties:
                        self:
                            type: string
                            example: 'https://sandbox.something.com/somelink'
...
        customer_request:
            type: object
            properties:
                email:
                    type: string
                    example: ahmad@mail.com
                password:
                    type: string
                    example: ahmadsecretpass
                name:
                    type: string
                    example: Ahmad
                dob:
                    type: string
                    example: 2001-08-09
                gender:
                    type: string
                    example: male
                address:
                    type: string
                    example: Dallas Westbrook ave.
                phone_number:
                    type: string
                    example: 082123123123
                avatar:
                    type: string
                    format: binary
        customer_response:
            type: object
            properties:
                id:
                    type: number
                    example: 11
                email:
                    type: string
                    example: ahmad@mail.com
                password:
                    type: string
                    example: ahmadsecretpass
                name:
                    type: string
                    example: Ahmad
                dob:
                    type: string
                    example: 2001-08-09
                gender:
                    type: string
                    example: male
                address:
                    type: string
                    example: Dallas Westbrook ave.
                phone_number:
                    type: string
                    example: 082123123123
                avatar:
                    type: string
                    example: https://source.unsplash.com/600x600/?random
                role:
                    type: string
                    example: customer
                created_at:
                    type: string
                    example: 2012-07-28T07:02:13:000+07:00
                updated_at:
                    type: string
                    example: 2012-07-28T07:02:13:000+07:00
        driver_request_create:
            type: object
            properties:
                email:
                    type: string
                    example: ahmad@mail.com
                password:
                    type: string
                    example: ahmadsecretpass
                name:
                    type: string
                    example: Ahmad
                dob:
                    type: string
                    example: 2001-08-09
                gender:
                    type: string
                    example: male
                address:
                    type: string
                    example: Dallas Westbrook ave.
                phone_number:
                    type: string
                    example: 082123123123
                avatar:
                    type: string
                    format: binary
                truck_type_id:
                    type: integer
                    example: 33
                ktp_file:
                    type: string
                    format: binary
                stnk_file:
                    type: string
                    format: binary
                driver_license_file:
                    type: string
                    format: binary
                age:
                    type: number
                    example: 22
                vehicle_identifier:
                    type: string
                    example: N 494 EY
                nik:
                    type: string
                    example: 33030012300004
                vehicle_picture:
                    type: string
                    format: binary
        driver_request_update:
            type: object
            properties:
                email:
                    type: string
                    example: ahmad@mail.com
                password:
                    type: string
                    example: ahmadsecretpass
                name:
                    type: string
                    example: Ahmad
                dob:
                    type: string
                    example: 2001-08-09
                gender:
                    type: string
                    example: male
                address:
                    type: string
                    example: Dallas Westbrook ave.
                phone_number:
                    type: string
                    example: 082123123123
                avatar:
                    type: string
                    format: binary
                age:
                    type: number
                    example: 22
                vehicle_identifier:
                    type: string
                    example: N 494 EY
                nik:
                    type: string
                    example: 33030012300004
...
        driver_response:
            type: object
            properties:
                email:
                    type: string
                    example: ahmad@mail.com
                password:
                    type: string
                    example: ahmadsecretpass
                name:
                    type: string
                    example: Ahmad
                dob:
                    type: string
                    example: 2001-08-09
                gender:
                    type: string
                    example: male
                address:
                    type: string
                    example: Dallas Westbrook ave.
                phone_number:
                    type: string
                    example: 082123123123
                avatar:
                    type: string
                    example: https://source.unsplash.com/600x600/?random
                role:
                    type: string
                    example: driver
                truck_type_id:
                    type: integer
                    example: 33
                truck_tupe:
                    $ref: '#/components/schemas/truck_type_response'
                ktp_file:
                    type: string
                    example: https://source.unsplash.com/600x600/?random
                stnk_file:
                    type: string
                    example: https://source.unsplash.com/600x600/?random
                driver_license_file:
                    type: string
                    example: https://source.unsplash.com/600x600/?random
                age:
                    type: number
                    example: 22
                vehicle_identifier:
                    type: string
                    example: N 494 EY
                nik:
                    type: string
                    example: 33030012300004
                vehicle_picture:
                    type: string
                    example: https://source.unsplash.com/600x600/?random
                status:
                    type: string
                    example: IDLE
                account_status:
                    type: string
                    example: PENDING
                created_at:
                    type: string
                    example: 2012-07-28T07:02:13:000+07:00
                updated_at:
                    type: string
                    example: 2012-07-28T07:02:13:000+07:00
...
        truck_type_response:
            type: object
            properties:
                id:
                    type: number
                    example: 77
                truck_type:
                    type: string
                    example: Mobil Box (A)
                max_volume:
                    type: number
                    example: 737
                max_weight:
                    type: number
                    example: 500
                price_per_distance:
                    type: number
                    example: 50000
                created_at:
                    type: number
                    example: 2012-07-28T07:02:13:000+07:00
                updated_at:
                    type: number
                    example: 2012-07-28T07:02:13:000+07:00
...
paths:
...
    /api/customers:
        post:
            description: Register a customer
            summary: Register a customer
            tags:
                - 'Auth & Registration'
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            $ref: '#/components/schemas/customer_request'
            responses:
                '200':
                    description: success registrating user
                    content:
                        application/json:
                            schema:
                                properties:
                                    status:
                                        type: string
                                        example: 'OK'
                                    code:
                                        type: number
                                        example: 200
                                    error:
                                        type: string
                                        example: null
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'https://sandbox.something.com/customers'
                                    data:
                                        $ref: '#/components/schemas/customer_response'
                                    pagination:
                                        $ref: '#/components/schemas/pagination'
                '422':
                    description: validation failure
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_validation_response'
                '500':
                    description: server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_server_response'
        put:
            description: Edit profile for authenticated user
            summary: Edit profile for authenticated user
            tags:
                - 'Customer'
            security:
                - JWTAuthCustomer: []
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            $ref: '#/components/schemas/customer_request'
            responses:
                '200':
                    description: success registrating user
                    content:
                        application/json:
                            schema:
                                properties:
                                    status:
                                        type: string
                                        example: 'OK'
                                    code:
                                        type: number
                                        example: 200
                                    error:
                                        type: string
                                        example: null
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'https://sandbox.something.com/customers'
                                    data:
                                        $ref: '#/components/schemas/customer_response'
                                    pagination:
                                        $ref: '#/components/schemas/pagination'
                '422':
                    description: validation failure
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_validation_response'
                '500':
                    description: server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_server_response'
        delete:
            tags:
                - 'Customer'
            description: Close/delete authenticated customer
            summary: Close/delete authenticated customer
            security:
                - JWTAuthCustomer: []
            responses:
                '200':
                    description: Success creating order's payment
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/success_response'
                '401':
                    description: Unauthorized user
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_unauthorized'
                '500':
                    description: Server Error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_server_response'
    # Registrasi Driver
    /api/drivers:
        post:
            description: Register a driver
            summary: Register a driver
            tags:
                - 'Auth & Registration'
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            $ref: '#/components/schemas/driver_request_create'
            responses:
                '200':
                    description: success registrating user
                    content:
                        application/json:
                            schema:
                                properties:
                                    status:
                                        type: string
                                        example: 'OK'
                                    error:
                                        type: string
                                        example: null
                                    code:
                                        type: number
                                        example: 200
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'https://sandbox.something.com/customers'
                                    data:
                                        $ref: '#/components/schemas/driver_response'
                                    pagination:
                                        $ref: '#/components/schemas/pagination'
                '422':
                    description: validation failure
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_validation_response'
                '500':
                    description: server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_server_response'
        put:
            description: Edit profile for authenticated driver
            summary: Edit profile for authenticated driver
            tags:
                - 'Driver'
            security:
                - JWTAuthDriver: []
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            $ref: '#/components/schemas/driver_request_update'
            responses:
                '200':
                    description: success registrating user
                    content:
                        application/json:
                            schema:
                                properties:
                                    status:
                                        type: string
                                        example: 'OK'
                                    code:
                                        type: number
                                        example: 200
                                    error:
                                        type: string
                                        example: null
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'https://sandbox.something.com/drivers'
                                    data:
                                        $ref: '#/components/schemas/driver_response'
                                    pagination:
                                        $ref: '#/components/schemas/pagination'
                '422':
                    description: validation failure
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_validation_response'
                '500':
                    description: server error
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error_server_response'
        get:
            description: Get list of drivers
            summary: Get list of drivers
            tags:
                - 'Admin'
            security:
                - JWTAuthAdmin: []
            parameters:
                - in: query
                  name: name
                  description: Search driver by name
                  schema:
                      type: string
                      example: Ahmad
                - in: query
                  name: gender
                  description: Filter drivers by gender
                  schema:
                      type: string
                      example: male
                - in: query
                  name: truck_type
                  description: Filter drivers by truck type
                  schema:
                      type: number
                      example: 27
                - in: query
                  name: sortName
                  description: Sort name alphabetically
                  schema:
                      type: boolean
                      example: 1
                - in: query
                  name: sortAge
                  description: Sort drivers by age
                  schema:
                      type: boolean
                      example: 1
            responses:
                '200':
                    description: Success getting list of drivers
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        example: 'OK'
                                    code:
                                        type: number
                                        example: 200
                                    error:
                                        type: string
                                        example: null
                                    links:
                                        type: object
                                        properties:
                                            self:
                                                type: string
                                                example: 'somelink'
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/driver_response'
                                    pagination:
                                        $ref: '#/components/schemas/pagination'

    # Auth login
...